Adding time out for procedures that hang when invoking. Changing collection of classes to write results directly instead of when done with all of the classes.
This commit includes a new time out function for fuzzing procedures. If a procedure hangs, it will now continue to the next after 5 seconds by default, modify using -timeout parameter. The collector now writes every COM class result directly to the output file instead of waiting for all to be ready and then finally write it to the output file.